[ADD, FIX] dados, ajuste skill
-Adicionado dados iniciais
-Feito ajustes em skill

+++ skillswap/src/main/java/br/com/skillswap/dto/skill/SkillRequestDTO.java
@@ -3,8 +3,8 @@ package br.com.skillswap.dto.skill;
 public class SkillRequestDTO {
     
 	private String nome;
-    private double tecAmp;
-    private double atkAdicional;
+    private int tecAmp;
+    private int atkAdicional;
     private double duracao;
     private double resfriamento;
     private String foto;
@@ -15,16 +15,16 @@ public class SkillRequestDTO {
 	public void setNome(String nome) {
 		this.nome = nome;
 	}
-	public double getTecAmp() {
+	public int getTecAmp() {
 		return tecAmp;
 	}
-	public void setTecAmp(double tecAmp) {
+	public void setTecAmp(int tecAmp) {
 		this.tecAmp = tecAmp;
 	}
-	public double getAtkAdicional() {
+	public int getAtkAdicional() {
 		return atkAdicional;
 	}
-	public void setAtkAdicional(double atkAdicional) {
+	public void setAtkAdicional(int atkAdicional) {
 		this.atkAdicional = atkAdicional;
 	}
