+++ skillswap/src/main/java/br/com/skillswap/modal/Skill.java
@@ -21,10 +21,10 @@ public class Skill {
     private String nome;
 
     @Column(nullable = false)
-    private double tecAmp;
+    private int tecAmp;
 
     @Column(nullable = false)
-    private double atkAdicional;
+    private int atkAdicional;
 
     @Column(nullable = false)
     private double duracao;
@@ -66,19 +66,19 @@ public class Skill {
         this.nome = nome;
     }
 
-    public double getTecAmp() {
+    public int getTecAmp() {
         return tecAmp;
     }
 
-    public void setTecAmp(double tecAmp) {
+    public void setTecAmp(int tecAmp) {
         this.tecAmp = tecAmp;
     }
 
-    public double getAtkAdicional() {
+    public int getAtkAdicional() {
         return atkAdicional;
     }
 
-    public void setAtkAdicional(double atkAdicional) {
+    public void setAtkAdicional(int atkAdicional) {
         this.atkAdicional = atkAdicional;
     }
